requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
opencv-python
scipy
torch
torchvision
transformers